requirements.txt
dash
dash-renderer
dash-core-components
dash-html-components
plotly
plotly-express

# testing...
gensim
nltk
requests
python-dotenv